// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/api/dashboard.service.ts
export interface GetDashboardRequestParams {
    version: string;
    date?: string;
    page?: number;
    limit?: number;
    locale?: string;
    trendDays?: number;
    timeZoneOffsetMinutes?: number;
    timeZoneId?: string;
}